:recycle: Refactor of routes into respective organized files
